Move PackageCopyBox
