+++ src/content/headsup/PackageCopyBox.tsx
